feat(authentication): add full authentication setup
Creates a new error type in "errors" called ForbiddenAccessError. Next, implements authentication.js
and authorizePermission.js in "middlware" and uses both middleware in userRoutes.js. It was
necessary to implement this setup now, since the user routes will need to pass the authentication
test before the user can interact with the backend.